requirements.txt
streamlit>=1.32.0
yfinance>=0.2.38
pandas>=2.0.0
plotly>=5.18.0
numpy>=1.24.0
pytz>=2024.1
groq>=0.4.0
gspread>=6.0.0
google-auth>=2.0.0
bcrypt>=4.0.0
opencc-python-reimplemented>=0.1.6
requests>=2.31.0
